requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
litellm
crewai
python-dotenv